# Dramatically reduce the time when rerun the doxygen with plantuml comments (Imporvement of performance)