@cheoljoo 2018-08-24
9.689 sec: Total elapsed time: 9.689 seconds
(of which 9.340 seconds waiting for external tools to finish)

According to upper timeline ,  generating images with java is a big burden.
So I try to remove java running process.   I guess we can reduce the java running time if we reuse already generated images.